services/object: Inline service checking node maintenance
Continues d29bbd0e64068ed761e0909db1ccc71c5998e62a.

+++ pkg/services/object/server.go
@@ -106,6 +106,10 @@ type FSChain interface {
 	// IsOwnPublicKey checks whether given pubKey assigned to Node in the NeoFS
 	// network map.
 	IsOwnPublicKey(pubKey []byte) bool
+
+	// LocalNodeUnderMaintenance checks whether local node is under maintenance
+	// according to the network map from FSChain.
+	LocalNodeUnderMaintenance() bool
 }
 
 // Storage groups ops of the node's object storage required to serve NeoFS API
@@ -202,6 +206,10 @@ func (s *server) Put(gStream protoobject.ObjectService_PutServer) error {
 			return err
 		}
 
+		if s.fsChain.LocalNodeUnderMaintenance() {
+			return s.sendStatusPutResponse(gStream, apistatus.ErrNodeUnderMaintenance)
+		}
+
 		if err = stream.Send(putReq); err != nil {
 			err = s.sendStatusPutResponse(gStream, err) // assign for defer
 			return err
@@ -233,6 +241,11 @@ func (s *server) Delete(ctx context.Context, req *protoobject.DeleteRequest) (*p
 	if err = signature.VerifyServiceMessage(delReq); err != nil {
 		return s.makeStatusDeleteResponse(err), nil
 	}
+
+	if s.fsChain.LocalNodeUnderMaintenance() {
+		return s.makeStatusDeleteResponse(apistatus.ErrNodeUnderMaintenance), nil
+	}
+
 	resp, err := s.srv.Delete(ctx, delReq)
 	if err != nil {
 		return s.makeStatusDeleteResponse(err), nil
@@ -265,6 +278,11 @@ func (s *server) Head(ctx context.Context, req *protoobject.HeadRequest) (*proto
 	if err = signature.VerifyServiceMessage(searchReq); err != nil {
 		return s.makeStatusHeadResponse(err), nil
 	}
+
+	if s.fsChain.LocalNodeUnderMaintenance() {
+		return s.makeStatusHeadResponse(apistatus.ErrNodeUnderMaintenance), nil
+	}
+
 	resp, err := s.srv.Head(ctx, searchReq)
 	if err != nil {
 		return s.makeStatusHeadResponse(err), nil
@@ -296,6 +314,11 @@ func (s *server) GetRangeHash(ctx context.Context, req *protoobject.GetRangeHash
 	if err = signature.VerifyServiceMessage(hashRngReq); err != nil {
 		return s.makeStatusHashResponse(err), nil
 	}
+
+	if s.fsChain.LocalNodeUnderMaintenance() {
+		return s.makeStatusHashResponse(apistatus.ErrNodeUnderMaintenance), nil
+	}
+
 	resp, err := s.srv.GetRangeHash(ctx, hashRngReq)
 	if err != nil {
 		return s.makeStatusHashResponse(err), nil
@@ -340,6 +363,11 @@ func (s *server) Get(req *protoobject.GetRequest, gStream protoobject.ObjectServ
 	if err = signature.VerifyServiceMessage(getReq); err != nil {
 		return s.sendStatusGetResponse(gStream, err)
 	}
+
+	if s.fsChain.LocalNodeUnderMaintenance() {
+		return s.sendStatusGetResponse(gStream, apistatus.ErrNodeUnderMaintenance)
+	}
+
 	err = s.srv.Get(
 		getReq,
 		&getStreamerV2{
@@ -385,6 +413,11 @@ func (s *server) GetRange(req *protoobject.GetRangeRequest, gStream protoobject.
 	if err = signature.VerifyServiceMessage(getRngReq); err != nil {
 		return s.sendStatusRangeResponse(gStream, err)
 	}
+
+	if s.fsChain.LocalNodeUnderMaintenance() {
+		return s.sendStatusRangeResponse(gStream, apistatus.ErrNodeUnderMaintenance)
+	}
+
 	err = s.srv.GetRange(
 		getRngReq,
 		&getRangeStreamerV2{
@@ -430,6 +463,11 @@ func (s *server) Search(req *protoobject.SearchRequest, gStream protoobject.Obje
 	if err = signature.VerifyServiceMessage(searchReq); err != nil {
 		return s.sendStatusSearchResponse(gStream, err)
 	}
+
+	if s.fsChain.LocalNodeUnderMaintenance() {
+		return s.sendStatusSearchResponse(gStream, apistatus.ErrNodeUnderMaintenance)
+	}
+
 	err = s.srv.Search(
 		searchReq,
 		&searchStreamerV2{
